Extract Prompt Studio highlight UI helpers
